Revert "add: new features"
This reverts commit e6cda5db3f0114a6e25f80d5b45c782d2ddd18dc.

+++ database/table.ts
@@ -55,9 +55,9 @@ const CreateTable = {
                 tx.executeSql(`CREATE TABLE IF NOT EXISTS sales (
                     id VARCHAR(50) PRIMARY KEY,
                     product_code BIGINT,
-                    price REAL,
+                    price BIGINT,
                     sold BIGINT,
-                    total REAL,
+                    total BIGINT,
                     created_by VARCHAR(30),
                     created_at VARCHAR(30))`)
             });
